+++ packages/swirl-components/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @getflip/swirl-components
 
+## 0.14.1
+
+### Patch Changes
+
+- [`5667bf7`](https://github.com/getflip/swirl/commit/5667bf7fb235a77cb9aa681929bdf2efb2670f2a)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Fix skeleton component width when
+  used inside flex containers
+
+## 0.14.0
+
+### Minor Changes
+
+- [#150](https://github.com/getflip/swirl/pull/150)
+  [`3afef39`](https://github.com/getflip/swirl/commit/3afef3969b0a7c675e50098dc74e7d3599ad8280)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add swirl-skeleton-box component
+
+* [#150](https://github.com/getflip/swirl/pull/150)
+  [`3afef39`](https://github.com/getflip/swirl/commit/3afef3969b0a7c675e50098dc74e7d3599ad8280)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Add swirl-skeleton-text component
+
 ## 0.13.0
 
 ### Minor Changes